# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "IoT Trap Monitoring System API",
        "description": "API для системы мониторинга IoT ловушек для борьбы с вредителями. Включает управление шлюзами, ловушками и регистрацию событий срабатывания.",
        "contact": {
            "email": "support@trap-monitoring.local"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:3000/api",
            "description": "Local Development Server"
        },
        {
            "url": "https://api.trap-monitoring.local/api",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/v1/gateways": {
            "get": {
                "tags": [
                    "Gateways"
                ],
                "summary": "Список шлюзов",
                "description": "Получить список всех IoT шлюзов с возможностью фильтрации по статусу",
                "operationId": "getGateways",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Фильтр по статусу шлюза",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "online",
                                "offline",
                                "maintenance"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/GatewayResource"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Gateways"
                ],
                "summary": "Создать шлюз",
                "description": "Создать новый IoT шлюз в системе",
                "operationId": "createGateway",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateGatewayRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Шлюз успешно создан",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GatewayResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The hardware id has already been taken."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "hardware_id": [
                                                    "The hardware id has already been taken."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/gateways/{id}": {
            "get": {
                "tags": [
                    "Gateways"
                ],
                "summary": "Детали шлюза",
                "description": "Получить детальную информацию об IoT шлюзе по его ID",
                "operationId": "getGatewayById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID шлюза",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GatewayResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Шлюз не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Gateway not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CreateGatewayRequest": {
                "required": [
                    "hardware_id",
                    "name"
                ],
                "properties": {
                    "hardware_id": {
                        "description": "Уникальный идентификатор оборудования шлюза",
                        "type": "string",
                        "maxLength": 255,
                        "example": "GW-001-ABC123"
                    },
                    "facility_id": {
                        "description": "UUID объекта, на котором устанавливается этот шлюз",
                        "type": "string",
                        "format": "uuid",
                        "example": "019bac36-54b6-70dc-94c3-cea624a51791"
                    },
                    "name": {
                        "description": "Название шлюза",
                        "type": "string",
                        "maxLength": 255,
                        "example": "Gateway Floor 1"
                    },
                    "status": {
                        "description": "Статус шлюза",
                        "type": "string",
                        "enum": [
                            "online",
                            "offline",
                            "maintenance"
                        ],
                        "example": "online",
                        "nullable": true
                    },
                    "location": {
                        "description": "Географическое расположение шлюза",
                        "properties": {
                            "latitude": {
                                "type": "number",
                                "format": "float",
                                "maximum": 90,
                                "minimum": -90,
                                "example": 50.4501
                            },
                            "longitude": {
                                "type": "number",
                                "format": "float",
                                "maximum": 180,
                                "minimum": -180,
                                "example": 30.5234
                            }
                        },
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "GatewayResource": {
                "properties": {
                    "id": {
                        "description": "UUID шлюза",
                        "type": "string",
                        "format": "uuid",
                        "example": "9d4e1c45-6b2a-4f8e-9c7d-3a1b5e8f4d2c"
                    },
                    "hardware_id": {
                        "description": "Идентификатор оборудования",
                        "type": "string",
                        "example": "GW-001-ABC123"
                    },
                    "facility_id": {
                        "description": "UUID объекта",
                        "type": "string",
                        "format": "uuid",
                        "example": "019bac36-54b6-70dc-94c3-cea624a51791"
                    },
                    "name": {
                        "description": "Название шлюза",
                        "type": "string",
                        "example": "Gateway Floor 1"
                    },
                    "location": {
                        "description": "Координаты расположения",
                        "properties": {
                            "latitude": {
                                "type": "number",
                                "format": "float",
                                "example": 50.4501
                            },
                            "longitude": {
                                "type": "number",
                                "format": "float",
                                "example": 30.5234
                            }
                        },
                        "type": "object"
                    },
                    "status": {
                        "description": "Статус шлюза",
                        "type": "string",
                        "enum": [
                            "online",
                            "offline",
                            "maintenance"
                        ],
                        "example": "online"
                    },
                    "is_online": {
                        "description": "Онлайн статус",
                        "type": "boolean",
                        "example": true
                    },
                    "last_seen_at": {
                        "description": "Время последней активности",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T12:00:00+00:00"
                    },
                    "created_at": {
                        "description": "Дата создания",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T10:00:00+00:00"
                    },
                    "updated_at": {
                        "description": "Дата обновления",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T12:00:00+00:00"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "HmacSignature": {
                "type": "apiKey",
                "description": "HMAC SHA-256 signature для защиты IoT endpoints. Используйте hash_hmac(\"sha256\", payload, gateway_api_key)",
                "name": "X-Signature",
                "in": "header"
            }
        }
    },
...
}